# openapi.yaml
openapi: 3.1.0
info:
  title: Ring Intercom Bridge
  version: 0.6.0
  description: >-
    Private control-plane and bounded WebRTC signaling contract for Ring
    Intercom Audio. The bridge never exposes vendor or bridge credentials.
servers:
  - url: http://127.0.0.1:8775
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Process liveness and delivery phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /metrics:
    get:
      operationId: prometheusMetrics
      responses:
        "200":
          description: Aggregate communication lifecycle and latency metrics
          content:
            text/plain:
              schema: {type: string}
  /v1/devices:
    get:
      operationId: listDevices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Alias-only device inventory
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Device"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/enrollments:
    post:
      operationId: startEnrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentStart"
      responses:
        "200":
          description: MFA challenge started or session completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentStarted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Another enrollment is active
        "422":
          description: Vendor credentials were rejected
        "429":
          description: Local or vendor rate limit
        "502":
          description: Vendor enrollment is unavailable
  /v1/enrollments/{enrollment}:
    parameters:
      - name: enrollment
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: verifyEnrollment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentVerify"
      responses:
        "200":
          description: Dedicated session persisted by the bridge
        "401":
          $ref: "#/components/responses/Unauthorized"
        "410":
          description: Challenge expired or was already consumed
        "422":
          description: Verification code was rejected
        "429":
          description: Vendor rate limit
        "502":
          description: Vendor enrollment is unavailable
    delete:
      operationId: cancelEnrollment
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Challenge absent or securely discarded
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/devices/{device}/capabilities:
    get:
      operationId: getCapabilities
      security:
        - bearerAuth: []
      parameters:
        - name: device
          in: path
          required: true
          schema:
            type: string
            pattern: "^[A-Za-z0-9_-]+$"
      responses:
        "200":
          description: Verified media capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown alias
  /v1/devices/{device}/audio/sessions:
    post:
      operationId: createAudioSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceAlias"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioSessionRequest"
      responses:
        "201":
          description: Bounded browser-to-Ring WebRTC session negotiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioSession"
        "400":
          description: Offer violates the audio-only PCMU contract
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown alias
        "409":
          description: This device already owns an active session
        "429":
          description: Per-device post-session cooldown
        "502":
          description: Ring signaling is unavailable
  /v1/devices/{device}/status:
    $ref: "./docs/openapi/control-paths.yaml#/Status"
  /v1/devices/{device}/unlock:
    $ref: "./docs/openapi/control-paths.yaml#/Unlock"
  /v1/devices/{device}/settings:
    $ref: "./docs/openapi/control-paths.yaml#/Settings"
  /v1/devices/{device}/audio/sessions/{session}:
    delete:
      operationId: deleteAudioSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceAlias"
        - name: session
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Local worker stopped, cooldown committed, or session already absent
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown alias
  /v1/devices/{device}/recordings:
    $ref: "./docs/openapi/recording-paths.yaml#/Recordings"
  /v1/devices/{device}/recordings/{recording}:
    $ref: "./docs/openapi/recording-paths.yaml#/Recording"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
  parameters:
    DeviceAlias:
      name: device
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9_-]+$"
  schemas:
    EnrollmentStart:
      $ref: "./docs/openapi/schemas.yaml#/EnrollmentStart"
    EnrollmentStarted:
      $ref: "./docs/openapi/schemas.yaml#/EnrollmentStarted"
    EnrollmentVerify:
      $ref: "./docs/openapi/schemas.yaml#/EnrollmentVerify"
    Health:
      $ref: "./docs/openapi/schemas.yaml#/Health"
    Device:
      $ref: "./docs/openapi/schemas.yaml#/Device"
    Capabilities:
      $ref: "./docs/openapi/schemas.yaml#/Capabilities"
    AudioSessionRequest:
      $ref: "./docs/openapi/schemas.yaml#/AudioSessionRequest"
    AudioSession:
      $ref: "./docs/openapi/schemas.yaml#/AudioSession"
    RecordingItem:
      $ref: "./docs/openapi/schemas.yaml#/RecordingItem"
    RecordingList:
      $ref: "./docs/openapi/schemas.yaml#/RecordingList"